==> jupyter_notebooks/1_automatically_process_messy_OCR_demo.ipynb <==
import pandas as pd
from csv import reader
import re
import os

DATA_DIR = 'data_sample'

filename = os.path.join(os.getcwd(), DATA_DIR, 'data.csv')

# Class & function definitions

# A list of numbers to parse as temperatures rather than as pKa values or remarks.
SPECIAL_TEMPS = ['14.3','31.2','41.4','50.3','22.2','10.5','31.5','24.9',
                 '10.7','7-9','25.7','24.2','75.6','20.4', '30-36','10-12',
                 '18-20','31.5','25-30','20-22','26.5','0.35','not_stated',
                 '19.5','9.6','22.5','25.5','9.3','10.3','48.8', '15.3', '25.3',
                 '35.2','25.1','24.6','34.9','13.1','20.1','13.2','19.6','25.9',
                 '30.5', '12.1','18.8']

# A regex string to match the number of a compound
FOUR_DIGIT_PATTERN = r'(?<!.)(?<!\d)( )*\d{4}(?!\d)[abc]?'

# Types of uncertainties that are assigned
UNCERT_TYPES = ('Uncert.','Uncert','Approx.','Approx','Rel','Rel.','V.uncert.','V.Uncert.')

# Different methods from the text. Currently the values are meaningless, and we just work with the indices.
METHOD_DICT = {'C1': 0,  'C2': 0, 'C3': 0,
             'E':0,'E1a': 0, 'E1b': 0, 'E1cg': 0, 'E1ch': 0, 'E1d': 0, 'E1e': 0,
             'E2a': 0, 'E2b': 0, 'E2c': 0, 
             'E3ag': 0, 'E3ah': 0, 'E3bg': 0, 'E3bh': 0, 'E3bq': 0, 'E3c': 0, 'E3d': 0, 'O5/E3bg':0, 'E3bg/O5' : 0,
             'O':0, 'O1': 0, 'O2': 0, 'O3': 0, 'O4': 0, 'O5': 0, 'O5a': 0, 'O6': 0, 'O7': 0, 
             'CAL': 0, 'CAT': 0, 'DIS': 0, 'FLU': 0, 'FLU/O5': 0, 'KIN': 0,
             'NMR': 0, 'RAM': 0, 'REF': 0, 'ROT': 0, 'SOL': 0,
             'R1a': 0, 'R1b': 0, 'R1a/b':0, 'R1c': 0, 'R1d': 0, 'R1e': 0,
             'R2a': 0, 'R2b': 0, 'R2c': 0,
             'R3a': 0, 'R3b': 0, 'R3c': 0, 'R3d': 0, 'R3f': 0, 'R3g': 0, 'R3h': 0,
             'R4': 0,
             'D1':0, 'D2':0, 'D3':0, 
             'K1a': 0, 'K1b':0, 'K1c': 0, 'K1d':0, 'K1e':0, 'K1f': 0,
             'K2a':0, 'K2b': 0, 'K2c':0, 'K2d':0, 'K2e':0}

class DataEntryBlock(object): 
    """
    A container that holds multiple individual data entries.
    The heirarchy is that a DataEntryBlock contains metadata across multiple entries, 
    such as the name of the molecule and the number in the reference book.
    """
    def __init__(self, number = None, name  = None, content=None, discarded = None, entries=None):
        if entries is None:
            entries = []
        if content is None:
            content = []
        self.number = number
        self.name = name
        self.content = content # A list of entries.
        self.discarded = discarded
        self.entries = entries
        
    def __str__(self): 
        return "# {}: {} | {} \nEntries: {}\nDiscarded: {}\n".format(self.number, self.name, self.content, len(self.entries), self.discarded)
    
class DataEntry(object):
    """
    A block of text partially parsed from one entry # to another
    This class provides a convenient way of storing and accessing the information.
    This contains data for the individual pKa entry, such as the pKa value and type, temperature, method, and reference.
    """
    def __init__(self, pkas=None, t=None, remarks = None, method=None, assessment=None, ref=None, subtables=None):
        if subtables is None:
            subtables = []
        if ref is None:
            ref = []
        if assessment is None:
            assessment = []
        if method is None:
            method = []
        if t is None:
            t = []
        if pkas is None:
            pkas = []
        self.pkas = pkas
        self.T = t
        self.remarks = remarks
        self.method = method
        self.assessment = assessment
        self.ref = ref
        self.subtables = subtables

    def __str__(self): 
        return "pKa: {} | T: {} | Remarks: {} | Method: {} | Assessment: {} | Ref: {} \n".format(self.pkas, self.T, self.remarks, self.method, self.assessment, self.ref)


def get_mol_score(word : str) -> int:
    """
    Takes a string and returns an integer score based on how many molecular formula-like tokens are in the string. 
    """
    elem_count = word.count('C') + word.count("\'") + word.count('H') + word.count('O') + word.count('F') + word.count('S') + word.count('N') + word.count('\"') + word.count('Â') + word.count('Ã') + word.count('°') + word.count('_') + 0.25*sum(c.isdigit() for c in word) + word.count('Br') + word.count('\"')
    elem_count = elem_count - word.count(',') - word.count('-') - word.count('(') - word.count(')') - word.count('=')
    return elem_count

def is_mol_formula(word : str) -> bool:
    """Tries to infer whether a continous string is a molecular formula string."""
    if get_mol_score(word) >= 2:
        return True
    else:
        return False

partial_replace_list = [("I $", "I ="),
                        (". .", "."),
                        ("-l-", "-1-"),
                        ("I = 0. ", "I = 0."),
                        ("I = 1. ", "I = 1."),
                        ("I - 0", "I = 0"),
                        (" a ", " = "),
                        ("K1103", "KNO3"),
                        (' = ', '='),
                        ('Rl', 'R1'),
                        ('/.Uncert', 'V.Uncert'),
                        ('El', 'E1'),
                        ('MaCl', 'NaCl'),
                        ('KN03', 'KNO3'),
                        ('KM03', 'KNO3'),
                        ('aCIOs', 'aClO4'),
                        (' 1 (', '1('),
                        (' 1(Na', '1(Na'),
                        (' 1(K', '1(K'),
                        ('KC1', 'KCl'),
                        ('Kla', 'K1a'),
                        ('KHO3', 'KNO3'),
                        ('V. Uncert', 'V. Uncert'),
                        ('NaC1', 'NaCl'),
                        ('Nactog', 'NaClO4'),
                        ('flactog', 'NaClO4'),
                        ('Blll', 'B111'),
                        ('Mlll', 'M111'),
                        ('Gll', 'G11'),
                        ('ybbcox', ''),
                        ('ybblox', ''),
                        ('yoblox', ''),
                        ('ybbLox', ''),
                        ('C10', 'ClO'),
                        ('Cl0', 'ClO'),
                        ('E3bg/05', 'E3bg/O5'),
                        ('N02', 'NO2'),
                        ('pky', 'pk1'),
                        (', p', ' p'),
                        ('07,R', 'O7,R'),
                        ('05,R', 'O5,R'),
                        (' (trans)', '(trans)'),
                        (' (cis)', '(cis)'),
                        ('(C0', '(CO'),
                        ('(Contd)', ''),
                        ('(contd)', ''),
                        ('pKg', 'pk1'),
                        ('pk,', 'pk1'),
                        ('Mll,', 'M11'),
                        ('Gll', 'G11'),
                        (' <', '<'),
                        (' >', '>'),
                        ('05a', 'O5a'),
                        ('thy1', 'thyl'),
                        ('oly1', 'olyl'),
                        ('E3pa', ''),
                        ('E2pa', ''),
                        ('E3po', ''),
                        ('opy1', 'opyl'),
                        ('eny1', 'enyl'),
                        ('exy1', 'exyl'),
                        ('ty1', 'tyl'),
                        ('ly1', 'lyl'),
                        ('ry1', 'ryl'),
                        ('ox0', 'oxo'),
                        ('0x', 'Ox'),
                        ('c 0', 'c=0'),
                        ('C 0', 'C=0'),
                        ('I 0', 'I=0'),
                        ('I  0', 'I=0'),
                        ('c  0', 'c=0'),
                        ('C  0', 'c=0'),
                        ('c104', 'ClO4'),
                        ('ciog', 'ClO4'),
                        ('1laC', 'NaC'),
                        ('cio4', 'ClO4'),
                        ('Nacion', 'NaClO4'),
                        ('i!', 'N'),
                        ('cioa', 'ClO4'),
                        ('(maclo', '(NaClO'),
                        ('ilacioq', 'NaClO4'),
                        ('ilaC1', 'NaCl'),
                        ('capa', ''),
                        ('pK of electron', 'p\'k of electron'),
                        ('(H20)', '(H2O)'),
                        ('(D20)', '(D2O)'),
                        (' (H2O)', '(D2O)'),
                        (' (D2O)', '(D2O)'),
                        (' (co)', '(co)'),
                        (' (cooh)', '(cooh)'),
                        (' (COOH)', '(COOH)'),
                        (' (NH)', '(NH)'),
                        (' (OH)', '(OH)'),
                        ('kno3', 'KNO3'),
                        ('onditions not stated', 'onditions-not-stated'),
                        ('not stated', 'not_stated'),
                        ('Not stated', 'not_stated') # 
]

full_replace_list = [("0","O"),
                     ("Tla","T1a"),
                     ("115","I15"),
                     ("120","I20"),
                    ]

def find_and_replace(row):
    """
    Cleans typos and artifacts from a data row (i.e. a list of strings).
    This is because OCR often makes the same type of mistakes.
    Returns a list of strings with cleaned strings.
    """
    for i, w in enumerate(row):
        for replacement in partial_replace_list:
            row[i] = row[i].replace(replacement[0], replacement[1])

        if row[i] in ['01','02','03','04','05','06','07','08','09', '010', '011', '012', '013', '014', '015', '016','017','018','019','020','021']:
            row[i] = w.replace('0','O')
        if row[i] in ['01 ','02 ','03 ','04 ','05 ','06 ','07 ','08 ','09 ', '010 ', '011 ', '012 ', '013 ', '014', '015 ', '016 ','017 ','018 ','019 ','020 ','021 ']:
            row[i] = w.replace('0','O')

        for replacement in full_replace_list:
            if row[i] == replacement[0]:
                row[i] = replacement[1]

    return row

def load_data(filename : str) -> list:
    """
    Takes a filename, assuming it is a csv, and returns the list of rows in the csv.
    """
    rows = []
    with open(filename, 'r',encoding="latin1") as read_obj:
        csv_reader = reader(read_obj)
        for _, row in enumerate(csv_reader):
            rows.append(row)
    return rows
        
def preprocess_data(data : list) -> list:    
    """
    Takes a list of blocks, and processes to clean up and standardize:
    (1) Removes redundant header rows.
    (2) Splits entry #s like '2002Propanoic' into '2002', 'Propanoic' separate cells
    (3) Checks for - and removes - molecular formulas, which are often improperly read into the sheet
    (4) Removes completely empty rows.
    This is all to make the data easier to parse into DataEntryBlock objects.
    Returns a list of blocks.
    
    """
    list_of_lists = []
    for indexline, row in enumerate(data): # Example of a row: ['2004', 'CH40 Methanol ', '', '', '', '', '', '', '']
        line_type = None # track whether we have moved onto a "new molecule" or are on the same entry.

        for index, item in enumerate(row): # Example of an item: "2004"
            row = find_and_replace(row)

            # Looping through words
            row[index] = item.strip()
            if index == 0: 
                # basically checking for a 4-digit number which indicates the entry number.
                # We want the 4-digit number to be stand-alone in the first column, with nothing else.
                match = re.search(FOUR_DIGIT_PATTERN,item)
                if match: 
                    # see if there are any nonmatches WITHIN that same string e.g. 'some2012test' --> ['some','test']
                    # If not, do nothing, we've done our job.
                    # If so, then move those nonmatches to after the 4-digit index, assume it's part of the name.
                    row[0] = '' 
                    nonmatches = re.split(FOUR_DIGIT_PATTERN, item)
                    nonmatches = [i for i in nonmatches if not (i in ['',' '])] # if no nonmatches, just get rid of it

                    if len(nonmatches) > 0:
                        nonmatches.reverse()
                        try:
                            row.insert(0, ''.join(nonmatches))
                        except TypeError:
                            pass
#                            print(f"Nonmatch @ : {nonmatches}")
                    row.insert(0,''.join(match.group()))
                    line_type = 'first'

        # Now we check for the molecular formula.
        for index, item in enumerate(row):
            if line_type == 'first':
                # Try to split up molecular formula from rest of string. Delete it if possible.
                split_string = item.split() # Ex: 'C2HO2F3 Acetic acid, trifluoro-' -> ['C2H02F3', 'Acetic', 'acid,', 'trifluoro-']

                if len(split_string) == 1: # this must go first for case of word like "CH2 Acetic acid"
                    if is_mol_formula(split_string[0]):
                        row.remove(row[index])

                # Delete anything that looks like a molecular formula.
                if len(split_string) > 1:
                    for word in split_string:
                        if is_mol_formula(word):
#                            molec_formula = word
                            split_string.remove(word)
                            fragment = ' '.join(split_string)
                            row[index] = fragment
                            break                

        # Remove row if it seems to contain a redundant header.
        if any(e in row for e in ['Remarks','No.','Molecular','formula','Method','Assessment', 'Ref','Ref.']) or all(e == '' for e in row): 
            print("Dropping linerow {}".format(indexline))
        else:
            list_of_lists.append(row)

    return list_of_lists

def process_data(data: list) -> list:
    """
    Turns a list of row data into a list of block datas. Assumes that entry #s are always in the first column.
    """
    block = DataEntryBlock(content = [], entries=[], name = None)
    block_list = []
    for _, row in enumerate(data):
        for index, item in enumerate(row):
            indices_to_del = []
            if index == 0:
            # Try to find a number of format "XXXX" in the first column.
            # If successful, split on the string if possible, and create a new block
            # Otherwise try to add it to the block.
                match = re.search(FOUR_DIGIT_PATTERN,item)
                if match is not None:
                    block_list.append(block) # Add previous block because it's done parsing
                    block = DataEntryBlock(number = match.group(), content = [], entries=[])
                    indices_to_del.append(0)

            # THE LAST THING TO BE EXECUTED: Delete all the indices we wanted to.
            if len(indices_to_del) > 0:
                for indice in indices_to_del:
                    del row[indice]
        
        block.content.append(row)

    block_list.append(block) # Add last block
    block_list.remove(block_list[0])# Remove first block
    return block_list

# Run the functions to process data into block form

data = load_data(filename)
for row in data:
    print(row)

# idea: As iterating through list, try to extract an entry # from first 3 columns
# Also create a DataEntry class and flag when there is a missing entry.
 
alldata = preprocess_data(data)
#alldata = preprocess_data(alldata) # Process twice 


it can now be seen that the data is "cleaner" in the sense that redundant rows are gone, and mol formulas are deleted.

for row in alldata:
    print(row)

blockdata = process_data(alldata)

for block in blockdata:
    print(block)

# Process from block form

def is_pka(value) -> bool:
    """
    Takes a value and tries to infer whether it should be considered a pKa (returning a bool).
    """
    pattern = re.compile("(\d){1,3}[.](\d){1,4}[ ]?(\([^\)]*\)){1}")
    try:
        if re.search(pattern, value) and 'I' not in value and 'k' not in value and 'K' not in value and 'Na' not in value and '?' not in value and 'L' not in value:
            return True
        float(value)
        if '.' in value and abs(float(value)) > 0.04 and not is_special_temp(value) and float(value) < 25.0: 
            return True
        else:
            return False
    except ValueError:
        return False

def is_temp(value) -> bool:
    """
    Takes a value and tries to infer whether it is a temperature, returning a bool.
    """
    try:
        int(value)
        if int(value) < 500:
            return True
        else:
            return False
    except ValueError:
        if is_special_temp(value):
            return True
        try:
            z = value.split('+')
            for a in z:
                int(a)
            return True
        except ValueError:
            return False

def is_special_temp(value) -> bool:
    """
    Takes a value and returns whether it is a 'special' temperature.
    These are unusually formatted temps, such as those in a range
    """
    pattern = re.compile("(\d){1,3}[+](\d){1,2}")
    pattern_q_mark = re.compile("(\d){1,3}(\([?]\))")
    if (value in SPECIAL_TEMPS) or re.search(pattern, value) or re.search(pattern_q_mark, value):
        return True
    else:
        return False

def find_name_from_edgecases(name, content):
    """
    Taking the edge-case IUPAC name and a row of data (content), try to see if the name can be inferred.
    Returns the processed name and leftover content
    """
    pattern = re.compile("(\d)?(\()?((\d\,)+)?([A-z]|\d)-[A-z\d\(\)[\]\-,]+") # e.g. matches 1,1-dimethyl; 2,2-mercapto-4-silica; 3,3,3-methyl-4-silica
    split_iupac_words = content.split()

    for ct,word in enumerate(split_iupac_words):
        try:
            matching = pattern.match(word).group()
            if name is None:
                name = matching
                split_iupac_words[ct] = split_iupac_words[ct].replace(matching, '', 1)
            elif name.contains(matching):
                pass
            else:
                name = name + matching
        except AttributeError:
            pass

    content = ' '.join(split_iupac_words)   

    return name, content

def extract_pkas(line, maxiter):
    """
    Given a line, try to find the pkas and corresponding pka types.
    Accepts a maxiter that tells how many columns right to go to.
    Outputs a dictionary that includes the compiled pKa info.
    """
    remaining_words = None

    pkas = []
    pkas_indices = []
    pkas_types = []
    pkas_types_indices = []
    pkas_dict = {}
    
    for index, content in enumerate(line):
        # CHECK 1: If word is in pkas.
        # Split word by spaces and then commas
        split_on = [' ', ',']
        split_text = line[index]
        for split in split_on:
            remaining_words = re.split(split,split_text)
            for ct,word in enumerate(remaining_words):
                if is_pka(word):
                    pkas.append(word)
                    pkas_indices.append([index,ct])
                    remaining_words[ct] = remaining_words[ct].replace(word, '',1)
                    if len(pkas) > 2: # If we're on a roll, this is probably a table
                        maxiter = maxiter + 1 
                elif ('pk' in word or 'pK' in word or 'Pk' in word or 'PK' in word) and ('pk=' not in word and '(pK' not in word and '(pk' not in word): 
                    if word in pkas_types:
                        pkas_types.append(word + '_' + str(index) + '_' + str(ct))
                    else:
                        pkas_types.append(word)
                    pkas_types_indices.append([index,ct])
                    remaining_words[ct] = remaining_words[ct].replace(word, '',1)
                elif split == ',' and ct > 0:
                    remaining_words[ct] = ',' + remaining_words[ct]
                elif split == ' ' and ct > 0:
                    remaining_words[ct] = ' ' + remaining_words[ct]
                    
            split_text = ''.join(remaining_words)
                    
        # Clear remaining words
        line[index] = ' '.join(remaining_words)
        
        if index == maxiter:
            break
        
    # Update pkas
    for count,val in enumerate(pkas):
        try:
            pkas_dict[pkas_types[count]] = val 
        except IndexError:
            pkas_dict['__'+str(count)] = val    

    return pkas_dict, line

def get_easy_values(entry, line, linenumber):
    """
    Try to find the "Easy" values given a line. 
    Fills out the entry's methods, temperatures, assessments, and corresponding indices,
    and returns this as another entry.
    """
    method_list = []
    method_indices = []
    ref_list = []
    ref_indices = []
    assessment_list = []
    t_list = []
    t_indices = []
    pattern = re.compile("^[A-NP-Z]{1}[0-9]{1,3}[a-z]{0,1}$")
    for index, content in enumerate(line):

        # CHECK 1: If word is in methods or is a temperature.
        # Split word by commas only for temperatures
        if (index <= 2) and linenumber == 0: # if we're early on in the data columns
            split_string = ' '
        else:
            split_string = '[, ]'
        remaining_words = re.split(split_string,line[index])
        for ct,word in enumerate(remaining_words):

            # Individually process the remaining text sequentially to see if we can extract "easy" regex matches.
            if remaining_words[ct] in list(METHOD_DICT.keys()):
                method_list.append(word)
                method_indices.append([index,ct])
                remaining_words[ct] = remaining_words[ct].replace(word, '')
            if is_temp(remaining_words[ct]):                        
                t_list.append(word)
                t_indices.append([index,ct])
                remaining_words[ct] = remaining_words[ct].replace(word, '')
            if remaining_words[ct] in UNCERT_TYPES: # search for uncertainty types
                assessment_list.append(word)
                remaining_words[ct] = remaining_words[ct].replace(word, '')

            # Search for refs
            try: 
                matching = pattern.match(remaining_words[ct]).group()
                if matching not in ['I','No','C','N','D20','H20'] and matching not in list(METHOD_DICT.keys()) and ' '.join(line).find("Other values") == -1:
                    remaining_words[ct] = remaining_words[ct].replace(matching, '')
                    ref_list.append(matching)
                    ref_indices.append([index, ct])
            except AttributeError:
                pass
            
        line[index] = ' '.join(remaining_words)            

    
    # Edge case: '05' improperly parsed as '5' and moved to T, and there are no methods, but there are other tags for the line.
    if len(t_list) < 3 and len(method_list) == 0 and (entry.assessment is not None or entry.ref is not None):
        numerals_to_check = ['01','02','03','04','05','06','07','08','09']
        nums_to_check = ['1','2','3','4','5','6','7','8','9']
        
        for numeral in nums_to_check:
            if numeral in t_list: # Can make this a function if this is a general problem
                method_list.append('O' + numeral)
                ind = t_list.index(numeral)
                t_list.remove(numeral)
                t_indices.remove(t_indices[ind])

        for numeral in numerals_to_check:
            if numeral in t_list: # Can make this a function if this is a general problem
                fixed_str = numeral.replace('0','O')
                method_list.append(fixed_str)
                ind = t_list.index(numeral)
                t_list.remove(numeral)
                t_indices.remove(t_indices[ind])

                
    # Update list with the new information that we've acquired.
    entry.method = method_list
    entry.method_indices = method_indices
    entry.T = t_list
    entry.t_indices = t_indices
    entry.ref = ref_list
    entry.ref_indices = ref_indices
    entry.assessment = assessment_list
    return entry, line

def populate_entries(blockdata: list) -> list:
    """
    This takes a list of block data and processes to identify temperatures, pKa values, pKa types, etc.
    Up to this point, the blockdata are just lumped together wihtout processing. This does the
    functional identification of the specific parameters.
    Returns the updated list of blockdata.
    """
    for block in blockdata:
        # Looping through the list of blocks data.

        for linenumber, line in enumerate(block.content):
            entry = DataEntry(pkas = [], t = [], remarks = None, method = [], ref = [], assessment = [], subtables = [])
            # Looping through lines in a single block
            
            # FIRST LOOP: Try to identify very specific edge cases because those are easiest to correct.
            # Edge case 1: "#,#-____" e.g. 2,2-dimethylphenyl e.g.
            for index, content in enumerate(line):
                block.name, content = find_name_from_edgecases(block.name, content)
                block.content[linenumber][index] = content   

                # Edge case 2: 'pk = ' or 'pK = ' in string
                pk_edge_cases = ['pK = ', 'pk = ']
                for case in pk_edge_cases:
                    block.content[linenumber][index] = block.content[linenumber][index].replace(case, 'pk=')

            # SECOND LOOP: Get "easy values" to correct a.k.a. uncert. types, methods, ref
            entry, line = get_easy_values(entry, line, linenumber)

            # update the 'max iter' for temp index
            try:
                maxiter = entry.t_indices[0][0] + entry.t_indices[0][1]
            except IndexError:
                maxiter = 4

            # THIRD LOOP: Get pKas. Only loop up to the temperatures index; failing this, to the fourth column, 
            # UNLESS there are many pKas, in which case this is probably a table.                        
            updated_pkas, line = extract_pkas(line, maxiter)            
            entry.pkas = updated_pkas
            
            # Penultimately, try to get name from first line
            if linenumber == 0:
                if block.name is None: 
                    block.name = ''
                block.name = block.name + ' '.join(line[0:3]) # do this addition in case of edge case detection earlier in loop
                line[0] = ''
                line[1] = ''
                line[2] = ''
                block.name = block.name.strip()
                block.name = block.name.lstrip('-')
            
            # Save the rest as remarks.
            entry.remarks = ' '.join(line).strip()
            
            if any([len(entry.pkas) > 0, len(entry.T) > 0, len(entry.remarks) > 0, len(entry.method) > 0, entry.assessment is not None, entry.ref is not None ] ):
                block.entries.append(entry)
            
            print(block.name)
            print(line)
            print(entry)

    return blockdata

blockdata = populate_entries(blockdata)
show_blockdata(blockdata)

def show_blockdata(blockdata) -> None :
    """ Helper function for visualizing the entries in the blockdata list. """
    for block in blockdata:
        print(block)
        for entry in block.entries:
            print(entry)
        print('\n')

show_blockdata(blockdata)

# Loop and get rid of blank rows
# Sort of post-processing step
def clean_empty_rows(blockdata):
    """
    Takes a block datum, cleans up artifacts (like extraneous trailing characters and missing rows), and returns the blockdata.
    """
    for blocknumber, block in enumerate(blockdata):

        # Also clear spaces in the word and some common mistakes.
        blockdata[blocknumber].name = re.sub(' +', ' ', block.name)
        blockdata[blocknumber].name = blockdata[blocknumber].name.rstrip(',')
        blockdata[blocknumber].name = blockdata[blocknumber].name.rstrip(' ')
        blockdata[blocknumber].name = blockdata[blocknumber].name.rstrip(',')
        blockdata[blocknumber].name = re.sub(' +', ' ', block.name)
        blockdata[blocknumber].name = blockdata[blocknumber].name.replace(' ,', ',')
        blockdata[blocknumber].name = blockdata[blocknumber].name.replace('- -', '-')
        blockdata[blocknumber].name = blockdata[blocknumber].name.replace(' - ', '-')
        blockdata[blocknumber].name = blockdata[blocknumber].name.replace(' -', '-')
        blockdata[blocknumber].name = blockdata[blocknumber].name.replace(' . ', '.')

        for count, entry in enumerate(block.entries):
            combined_data = [entry.pkas, entry.T, [entry.remarks], entry.method, entry.assessment, entry.ref]
            unique_values = list(set([item for sublist in combined_data for item in sublist]))

            # (1) Eliminate empty rows
            if len(unique_values) == 1 and unique_values[0] == '':
                blockdata[blocknumber].entries.remove(entry)
            
        for count, entry in enumerate(block.entries): # separate loop in case stuff got deleted
            # (2) Clean up a bit
            blockdata[blocknumber].entries[count].remarks = blockdata[blocknumber].entries[count].remarks.replace(' . ', '.')
            blockdata[blocknumber].entries[count].remarks = blockdata[blocknumber].entries[count].remarks.replace(' .', '.')

    return blockdata

blockdata = clean_empty_rows(blockdata)

def decompose_pkas(pka_dict):
    """
    Helper function that decomposes a pKa dictionary into its key and value items individually.
    """
    pka_types = list(pka_dict.keys())
    pkas = list(pka_dict.values())
    return pka_types, pkas

## Post-post processing: Operating 'in the cleaned entries'

# Subtables and post-processing of remarks
EDGE_CASE_REMARKS = ['I -', 'I =', 'I=','C=','C =', 'C  ', 'I ', 'm = '
                     'Concentration', 'Calculated','calculated', 'conditions',
                     'Conditions', 'aq.', 'Room temp', 'Mixed ', 'mixed ', 'not stated',
                     'Compound', 'No detail', "not stated", "m=","I=",'c=', 'C=', 'c =', 'Values', 'other',
                     'Probably', 'C<', 'C <', 'I<', 'I <', ' not ', ' in 0', 'room temp']

def process_multi_temp(entry):
    """
    Take an entry and clean up its temperatures. 
    Return the cleaned-up temperatures in a new entry. 
    """
    while len(entry.T) > 1:
    
        # Case 0: if there's "not_stated", then that's it
        for temp in entry.T:
            if temp == 'not_stated':
                entry.T.remove('not_stated')
                entry.remarks = entry.remarks + ' '.join(entry.T)
                entry.T = ['not_stated']
                break
        
        try:
            z = [int(float(k)%3) for k in entry.T]
        except SystemError:
            z = []
    
        # Case 1: pKa not present
        if len(entry.pkas) == 0 and len(entry.T) > 1:
            # check if there is a 'special' T. If so, move this to the pKa
            for temp in entry.T:
                if is_special_temp(temp):
                    entry.pkas['__0c'] = temp
                    entry.T.remove(temp)
    
        # Case 2: Method not present
        if len(entry.method) == 0 and len(entry.T) > 1:
            for temp in entry.T:
                try:
                    if 0 < int(temp) < 10:
                        entry.method.append('O' + str(int(temp)))
                        entry.T.remove(temp)
                        break
                except ValueError:
                    pass
        
        # Case 3: negative temp is a remarks thing
        if ('-1' in entry.T or '-2' in entry.T) and len(entry.T) > 1:
            for temp in entry.T:
                if temp == "-1" or temp == "-2":
                    entry.remarks = entry.remarks + "temp"
                    entry.T.remove(temp)
                    break
    
        # Case 4: there's a "25" in the
        if '25' in entry.T and len(entry.T) > 1:
            entry.T.remove('25')
            entry.remarks = entry.remarks + ' '.join(entry.T)
            entry.T = ['25']
    
        # Case 5: just take the lowest %5 one
        if 0 in z and len(entry.T) > 1:
            epsilon = 0.001
            for temp in entry.T:
                if float(temp) % 5 < epsilon:
                    entry.T.remove(temp)
                    entry.remarks = entry.remarks + ' '.join(entry.T)
                    entry.T = [temp]
                    break
    
        # Case 6: just take the 1st temperature found
        if len(entry.T) > 1:
            entry.remarks = entry.remarks + ' '.join(entry.T[1:])
            entry.T = [entry.T[0]]

    return entry
    
def postpost_process(blockdata):
    """
    Final postprocessing step that further cleans up the datatable.
    Identifies subtables (which are tables in data entries that contain pKa data over different
    pressures, temperatures, etc) and saves those as a separate parameter.
    Also moves some edge case info from the names to the remarks. And try to process 
    if multiple temperatures were inferred from previous steps.
    """
    for block in blockdata:
        if len(block.entries) == 0:
            print("Skipping")
            continue
        assume_subtable = False
        subtable_type = None

        # (1) Identify subtables
        for count, entry in enumerate(block.entries):
            first_loop = None
            combined_data = [entry.pkas, entry.T, [entry.remarks], entry.method, entry.assessment, entry.ref]
            unique_values = list(set([item for sublist in combined_data for item in sublist]))
            subtable_index = None

            # Try to infer whether a subtable is present
            if ("Variation " in entry.remarks and "with" in entry.remarks) or "with pressure" in entry.remarks:
                if "with pressure" in entry.remarks:
                    subtable_type = "pressure"
                assume_subtable = True
                if count > 0:
                    subtable_index = count-1
                else:
                    subtable_index = count
                block.entries[subtable_index].subtables.append(entry.remarks)
                block.entries[count].remarks = ''
                first_loop = True

            if assume_subtable: 
                added_pressure = False
                if len(entry.pkas) > 2: # Long list of pkas 
                    block.entries[subtable_index].subtables.append(entry.pkas)
                    block.entries[count].pkas = {}
                if len(entry.T) > 2: # Long list of temps
                    block.entries[subtable_index].subtables.append(entry.T)
                    block.entries[count].T = []

                numbers = sum(c.isdigit() for c in entry.remarks)
                letters = sum(c.isalpha() for c in entry.remarks)

                # ... if the text includes some indication of pressure
                if ("(atm)" in entry.remarks or "(bar)" in entry.remarks) or (subtable_type == 'pressure' and numbers > 3 and '=' not in entry.remarks and numbers > letters): 
                    block.entries[subtable_index].subtables.append(entry.remarks)
                    block.entries[count].remarks = ''
                    added_pressure = True
                if len(entry.T) <= 2 and len(entry.pkas) <= 2 and added_pressure == False and first_loop == False and len(unique_values) <= 1:
                    assume_subtable = False    
                    subtable_type = None
                first_loop = False

        # (2) Move certain edge cases from name to remarks if detected
        case_list = []
        for case in EDGE_CASE_REMARKS: 
            ind_find = block.name.find(case)
            if ind_find != -1:
                case_list.append(ind_find)
        if len(case_list) > 0:
            name_str = block.name[0:min(case_list)]
            remarks_str = block.name[min(case_list):]
            block.name = name_str
            if block.entries[0].remarks == '':
                block.entries[0].remarks = remarks_str
            else:
                block.entries[0].remarks = block.entries[0].remarks + '\n' + remarks_str

        # (3) Edge case if '=' in remarks and num. pka types > pk, then move it from the remarks to the pKas
        for entry in block.entries:
            matching_indices = [m.start() for m in re.finditer('=', entry.remarks)]
            if len(matching_indices) > 0:
                pka_types, pkas = decompose_pkas(entry.pkas)
                for i, pka_type in reversed(list(enumerate(pka_types))):
                    if '__' in pka_type and len(pka_types) < len(pkas):
                        pka_types.remove(pka_type)
                        del pkas[i]
                        entry.pkas.pop(pka_type)
                    for j, pka in reversed(list(enumerate(pkas[len(pkas):]))):
                        res = list(entry.remarks)
                        try:
                            res.insert(matching_indices[j], pka)
                        except IndexError:
                            res.insert(matching_indices[len(matching_indices)-1], pka)
                        res = ''.join(res)

        # (4) Handle multiple temperatures
        for entry in block.entries:
            entry = process_multi_temp(entry)

        # (5) Add remarks of near-empty rows to previous rows
        for count, entry in reversed(list(enumerate(block.entries))):
            if sum([len(entry.pkas), len(entry.T), len(entry.method), len(entry.assessment), len(entry.ref)]) == 0 and entry.remarks != '':
                if count > 0:
                    if block.entries[count-1].remarks == '':
                        block.entries[count-1].remarks = entry.remarks
                    else:
                        block.entries[count-1].remarks = block.entries[count-1].remarks + '\n' + entry.remarks
                elif count == 0:
                    block.name = block.name + ' ' + entry.remarks 
                
                block.entries[count].remarks = ''
                
        # (6) Another edge case. What if there's no pKa, but 1 special temperature?
        # Move it from temp to pKa
        for entry in block.entries:
            if len(entry.pkas) == 0 and len(entry.T) == 1:
                if is_special_temp(entry.T[0]):
                    entry.pkas['__0'] = entry.T[0]
                    entry.T.remove(entry.T[0])
                    

    return blockdata

blockdata = postpost_process(blockdata)
blockdata = clean_empty_rows(blockdata)

show_blockdata(blockdata)

# Create a new CSV

list_of_dicts = []

df = pd.DataFrame(columns=['Entry #', 'IUPAC', 'pKa type','pKa','T','Remarks','Method','Assessment','Ref', '(Subtable)'])

num = 0
# we would prefer to do this dict-by-dict, but this is OK in a pinch
for block in blockdata:
    l = [None]*10
    l[0] = block.number
    l[1] = block.name
    df.loc[num] = l
    for entry in block.entries:
        l[0] = None
        l[1] = None
        l[2] = ', '.join(list(entry.pkas.keys()))
        l[3] = ', '.join(list(entry.pkas.values()))
        l[4] = ', '.join(entry.T)
        l[5] = entry.remarks
        l[6] = ', '.join(entry.method)
        l[7] = ', '.join(entry.assessment)
        try:
            l[8] = ', '.join(entry.ref)
        except SystemError:
            l[8] = ''
        if entry.subtables:
            l[9] = entry.subtables
        else:
            l[9] = ''
        num = num + 1
        df.loc[num] = l
    num = num + 1

df.head(999)

filename_out = "test_sample_out.csv"
df.to_csv(filename_out,index=False)
